# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TEST OpenApi  Документація",
        "description": "L5 Swagger OpenApi опис",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/customerorders": {
            "get": {
                "tags": [
                    "Customer orders"
                ],
                "summary": "Сторінка клієнтів",
                "description": "Повертає json список клієнтів",
                "operationId": "07af356236954a0a51db36f57fa4d8f2",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/customerorders/store": {
            "post": {
                "tags": [
                    "Customer orders"
                ],
                "summary": "Створити нового клієнта",
                "description": "Повертає статус",
                "operationId": "7eb5860ab7bd1e4d532d8ceb35e91150",
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "status_id": {
                                        "type": "interger",
                                        "example": "1"
                                    },
                                    "delivery_type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "payment_type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "sum": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "2000.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/customerorders/show/{id}": {
            "get": {
                "tags": [
                    "Customer orders"
                ],
                "summary": "Дані одного клієнта",
                "description": "Повертає json",
                "operationId": "a7be5539cd9150731957d86d914f6b7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/customerorders/update/{id}": {
            "put": {
                "tags": [
                    "Customer orders"
                ],
                "summary": "Зміна даних клієнта",
                "description": "Повертає статус",
                "operationId": "8a999c00c7c71506fa8a1f8bfd70ca62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id акаунта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "status_id": {
                                        "type": "interger",
                                        "example": "1"
                                    },
                                    "delivery_type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "payment_type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "sum": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "2000.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/customerorders/destroy/{id}": {
            "delete": {
                "tags": [
                    "Customer orders"
                ],
                "summary": "Видалити клієнта",
                "description": "Повертає статус",
                "operationId": "e8bece2f62dfca80c091095270406dd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}